# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: responses
    description: Public questionnaire submissions
  - name: admin
    description: Private validation dashboard
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /responses:
    post:
      operationId: createResponse
      tags: [responses]
      summary: Save a completed commuter questionnaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponseInput"
      responses:
        "201":
          description: Saved response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseCreated"
        "400":
          description: Invalid questionnaire data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events:
    post:
      operationId: createEvent
      tags: [responses]
      summary: Record an anonymous questionnaire event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "204":
          description: Event recorded
        "400":
          description: Invalid event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get grouped questionnaire results
      parameters:
        - $ref: "#/components/parameters/AdminPassword"
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
      responses:
        "200":
          description: Results summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
        "401":
          description: Invalid admin password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/responses:
    get:
      operationId: getAdminResponses
      tags: [admin]
      summary: List submitted questionnaire responses
      parameters:
        - $ref: "#/components/parameters/AdminPassword"
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
      responses:
        "200":
          description: Raw responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminResponse"
        "401":
          description: Invalid admin password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/export.csv:
    get:
      operationId: exportAdminResponses
      tags: [admin]
      summary: Download questionnaire responses as CSV
      parameters:
        - $ref: "#/components/parameters/AdminPassword"
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Invalid admin password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/posters/{posterId}:
    patch:
      operationId: updateAdminPoster
      tags: [admin]
      summary: Update a poster registry entry
      parameters:
        - $ref: "#/components/parameters/AdminPassword"
        - name: posterId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/PosterSource"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PosterRegistryUpdate"
      responses:
        "200":
          description: Updated poster registry entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PosterRegistryItem"
        "400":
          description: Invalid poster registry data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid admin password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    AdminPassword:
      name: X-Admin-Password
      in: header
      required: true
      schema:
        type: string
    FromDate:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date
    ToDate:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date
  schemas:
    PosterSource:
      type: string
      enum: [P01, P02, P03, direct_unknown]
    PosterRegistryUpdate:
      type: object
      required: [locationName, latitude, longitude]
      properties:
        locationName:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
    PosterRegistryItem:
      type: object
      required: [posterId, locationName, latitude, longitude]
      properties:
        posterId:
          $ref: "#/components/schemas/PosterSource"
        locationName:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
    AttributionSummary:
      allOf:
        - $ref: "#/components/schemas/PosterRegistryItem"
        - type: object
          required: [label, landingVisits, estimatedUniqueVisitors, surveyStarts, completedSurveys, completionRate]
          properties:
            label:
              type: string
            landingVisits:
              type: number
            estimatedUniqueVisitors:
              type: number
            surveyStarts:
              type: number
            completedSurveys:
              type: number
            completionRate:
              type: number
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ScheduleDayInput:
      type: object
      required: [day, active, arrival, departure]
      properties:
        day:
          type: string
          enum: [monday, tuesday, wednesday, thursday, friday]
        active:
          type: boolean
        arrival:
          type: string
          minLength: 1
        departure:
          type: string
          minLength: 1
    ResponseInput:
      type: object
      required:
        - role
        - posterSource
        - submissionId
        - livesInSageCreek
        - isUofMStudent
        - schedule
        - arrivalFlexibility
        - departureFlexibility
        - rideDirection
        - scheduleChangeFrequency
        - dealbreaker
        - intentLevel
        - email
        - phone
        - prefersText
      properties:
        role:
          type: string
          enum: [driver, rider]
        posterSource:
          $ref: "#/components/schemas/PosterSource"
        submissionId:
          type: string
          minLength: 1
        livesInSageCreek:
          type: boolean
        isUofMStudent:
          type: boolean
        schedule:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ScheduleDayInput"
        arrivalFlexibility:
          type: string
          minLength: 1
        departureFlexibility:
          type: string
          minLength: 1
        rideDirection:
          type: string
          minLength: 1
        maxDetour:
          type: ["string", "null"]
        seats:
          type: ["string", "null"]
        maxPickupWalk:
          type: ["string", "null"]
        currentTransportMethod:
          type: ["string", "null"]
        currentCommuteDuration:
          type: ["string", "null"]
        minimumMonthlyCompensation:
          type: ["string", "null"]
        maximumMonthlyWillingnessToPay:
          type: ["string", "null"]
        scheduleChangeFrequency:
          type: string
          minLength: 1
        dealbreaker:
          type: string
          minLength: 1
        dealbreakerOther:
          type: ["string", "null"]
        intentLevel:
          type: string
          minLength: 1
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        prefersText:
          type: boolean
        utmSource:
          type: ["string", "null"]
        utmMedium:
          type: ["string", "null"]
        utmCampaign:
          type: ["string", "null"]
        referrer:
          type: ["string", "null"]
    ResponseCreated:
      type: object
      required: [id, createdAt]
      properties:
        id:
          type: number
        createdAt:
          type: string
          format: date-time
    EventInput:
      type: object
      required: [eventName, posterSource, anonymousVisitorId, browserSessionId]
      properties:
        eventName:
          type: string
          enum:
            - landing_viewed
            - driver_role_selected
            - rider_role_selected
            - step_reached
            - form_completed
            - survey_started
        posterSource:
          $ref: "#/components/schemas/PosterSource"
        anonymousVisitorId:
          type: string
          minLength: 1
        browserSessionId:
          type: string
          minLength: 1
        role:
          type: ["string", "null"]
        step:
          type: ["number", "null"]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AdminResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseInput"
        - type: object
          required: [id, createdAt]
          properties:
            id:
              type: number
            createdAt:
              type: string
              format: date-time
    DistributionItem:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
        count:
          type: number
    ScheduleBucket:
      type: object
      required: [day, time, drivers, riders]
      properties:
        day:
          type: string
        time:
          type: string
        drivers:
          type: number
        riders:
          type: number
    AdminSummary:
      type: object
      required:
        - total
        - drivers
        - riders
        - interestedDrivers
        - interestedRiders
        - driverCompensation
        - riderWillingness
        - weekdayActivity
        - arrivalDistribution
        - reliability
        - driverDealbreakers
        - riderDealbreakers
        - transportMethods
        - commuteDurations
        - potentialOverlap
        - attribution
        - posterRegistry
      properties:
        total:
          type: number
        drivers:
          type: number
        riders:
          type: number
        interestedDrivers:
          type: number
        interestedRiders:
          type: number
        driverCompensation:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        riderWillingness:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        weekdayActivity:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        arrivalDistribution:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        reliability:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        driverDealbreakers:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        riderDealbreakers:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        transportMethods:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        commuteDurations:
          type: array
          items:
            $ref: "#/components/schemas/DistributionItem"
        potentialOverlap:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleBucket"
        attribution:
          type: array
          items:
            $ref: "#/components/schemas/AttributionSummary"
        posterRegistry:
          type: array
          items:
            $ref: "#/components/schemas/PosterRegistryItem"