# lib/api-zod/src/generated/types/adminSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttributionSummary } from './attributionSummary';
import type { DistributionItem } from './distributionItem';
import type { PosterRegistryItem } from './posterRegistryItem';
import type { ScheduleBucket } from './scheduleBucket';

export interface AdminSummary {
  total: number;
  drivers: number;
  riders: number;
  interestedDrivers: number;
  interestedRiders: number;
  driverCompensation: DistributionItem[];
  riderWillingness: DistributionItem[];
  weekdayActivity: DistributionItem[];
  arrivalDistribution: DistributionItem[];
  reliability: DistributionItem[];
  driverDealbreakers: DistributionItem[];
  riderDealbreakers: DistributionItem[];
  transportMethods: DistributionItem[];
  commuteDurations: DistributionItem[];
  potentialOverlap: ScheduleBucket[];
  attribution: AttributionSummary[];
  posterRegistry: PosterRegistryItem[];
}