# lib/api-zod/src/generated/types/attributionSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PosterRegistryItem } from './posterRegistryItem';

export type AttributionSummary = PosterRegistryItem & {
  label: string;
  landingVisits: number;
  estimatedUniqueVisitors: number;
  surveyStarts: number;
  completedSurveys: number;
  completionRate: number;
};