# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EventInputEventName } from './eventInputEventName';
import type { PosterSource } from './posterSource';

export interface EventInput {
  eventName: EventInputEventName;
  posterSource: PosterSource;
  /** @minLength 1 */
  anonymousVisitorId: string;
  /** @minLength 1 */
  browserSessionId: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  step?: number | null;
}