# lib/api-zod/src/generated/types/posterRegistryItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PosterSource } from './posterSource';

export interface PosterRegistryItem {
  posterId: PosterSource;
  /** @nullable */
  locationName: string | null;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
}