# lib/api-zod/src/generated/types/posterRegistryUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PosterRegistryUpdate {
  /** @nullable */
  locationName: string | null;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
}