# lib/api-zod/src/generated/types/posterSource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PosterSource = typeof PosterSource[keyof typeof PosterSource];


export const PosterSource = {
  P01: 'P01',
  P02: 'P02',
  P03: 'P03',
  direct_unknown: 'direct_unknown',
} as const;