# lib/api-zod/src/generated/types/responseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PosterSource } from './posterSource';
import type { ResponseInputRole } from './responseInputRole';
import type { ScheduleDayInput } from './scheduleDayInput';

export interface ResponseInput {
  role: ResponseInputRole;
  posterSource: PosterSource;
  /** @minLength 1 */
  submissionId: string;
  livesInSageCreek: boolean;
  isUofMStudent: boolean;
  /** @minItems 1 */
  schedule: ScheduleDayInput[];
  /** @minLength 1 */
  arrivalFlexibility: string;
  /** @minLength 1 */
  departureFlexibility: string;
  /** @minLength 1 */
  rideDirection: string;
  /** @nullable */
  maxDetour?: string | null;
  /** @nullable */
  seats?: string | null;
  /** @nullable */
  maxPickupWalk?: string | null;
  /** @nullable */
  currentTransportMethod?: string | null;
  /** @nullable */
  currentCommuteDuration?: string | null;
  /** @nullable */
  minimumMonthlyCompensation?: string | null;
  /** @nullable */
  maximumMonthlyWillingnessToPay?: string | null;
  /** @minLength 1 */
  scheduleChangeFrequency: string;
  /** @minLength 1 */
  dealbreaker: string;
  /** @nullable */
  dealbreakerOther?: string | null;
  /** @minLength 1 */
  intentLevel: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  prefersText: boolean;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  /** @nullable */
  referrer?: string | null;
}